rename perforceModel to perforceApi
